Add GitHub workflow for DockerHub deployment (#8)
### What
Add GitHub workflow to build and push Docker images to DockerHub on main
branch commits. Move Dockerfile to repository root and remove
Makefile-based Docker build process.

  ### Why
Automates Docker image publishing via DockerHub. Moved the Dockerfile to
the root as this is common to find a Dockerfile in the root of small
simple projects. Removed the Makefile with the docker commands because
some of the commands, using sudo, are not needed and the indirection is
unnecessary.